+++ gradle.properties
@@ -1,2 +1,8 @@
-group=com.marklogic
-version=1.1-SNAPSHOT
+# Define these on the command line to publish to OSSRH
+# See https://central.sonatype.org/publish/publish-gradle/#credentials for more information
+mavenCentralUsername=
+mavenCentralPassword=
+mavenCentralUrl=https://oss.sonatype.org/service/local/staging/deploy/maven2/
+#signing.keyId=YourKeyId
+#signing.password=YourPublicKeyPassword
+#signing.secretKeyRingFile=PathToYourKeyRingFile
