+++ gradle/wrapper/gradle-wrapper.properties
@@ -3,4 +3,5 @@ distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.5.1-bin.zip
+# Using 6.9 until ml-gradle 4.3 is available, which will support 7.x
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.9-bin.zip
