+++ marklogic-unit-test-client/gradle.properties
@@ -1,10 +1,10 @@
 # These properties are used for the test application that ml-gradle deploys so we can test out marklogic-unit-test
 mlAppName=marklogic-unit-test-testapp
 mlHost=localhost
-mlGradleVersion=4.1.1
 mlUsername=admin
 mlPassword=admin
-mlRestPort=8090
+mlRestPort=8008
+mlContentForestsPerHost=1
 
 # Load both the marklogic-unit-test framework code and some test modules
 mlModulePaths=../marklogic-unit-test-modules/src/main/ml-modules,src/test/ml-modules
